On changing fit type, enable/disable sequential fit option
If not a "single" fit, it doesn't make sense to have a sequential fit.
Add an option to allow/disallow this independently of the regular "Fit" button.
(If no function is set, it will still be disabled whether or not the fit type is "single").

Handle user changing "runs" box - set fit type if appropriate.

re #15518