Fix bug where getting runs got stuck
Because getRuns() re-ran the file search every time, it caused userChangedRuns
to get called after (when files were found). This called getRuns() and it started
again...

Now getRuns() is just a getter and doesn't start the search again.
The places it's called all happen after the filesFound signal so there's
no need to run the search again.

re #15518